# feat(wal): reduce concurrent conflicts between block write operations and poll operations