--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/block/Block.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/block/Block.java
@@ -29,6 +29,11 @@ public interface Block {
     long startOffset();
 
     /**
+     * The start time of this block*
+     * @return
+     */
+    long startTime();
+    /**
      * Append a record to this block.
      * Cannot be called after {@link #data()} is called.
      *