--- a/s3stream/src/main/java/com/automq/stream/s3/wal/impl/block/BlockBatch.java
+++ b/s3stream/src/main/java/com/automq/stream/s3/wal/impl/block/BlockBatch.java
@@ -21,6 +21,7 @@ public class BlockBatch {
     private final Collection<Block> blocks;
     private final long startOffset;
     private final long endOffset;
+    private final long blockBatchSize;
 
     public BlockBatch(Collection<Block> blocks) {
         assert !blocks.isEmpty();
@@ -33,6 +34,9 @@ public class BlockBatch {
             .map(b -> b.startOffset() + b.size())
             .max(Long::compareTo)
             .orElseThrow();
+        this.blockBatchSize = blocks.stream()
+            .mapToLong(Block::size)
+            .sum();
     }
 
     public long startOffset() {
@@ -45,6 +49,10 @@ public class BlockBatch {
 
     public Collection<Block> blocks() {
         return Collections.unmodifiableCollection(blocks);
+    }
+
+    public long blockBatchSize(){
+        return blockBatchSize;
     }
 
     public Iterator<CompletableFuture<AppendResult.CallbackResult>> futures() {