[issue](https://github.com/AutoMQ/automq/issues/1550)