@RapperCL 2024-07-11
> It seems that the "lock separation" mentioned in #1550 is not included in this PR. Is it still in progress?

I additionally added a `pollBlockLock` to handle the concurrency safety of poll operations, which is in this PR.