@RapperCL 2024-07-11
> We have provided a tool called "[WriteBench](https://github.com/AutoMQ/automq/blob/main/s3stream/src/main/java/com/automq/stream/s3/wal/benchmark/WriteBench.java)" for testing WAL performance. You can use it to compare the performance before and after the modifications to confirm their effectiveness.

Okay, I'll find some time later to test it.